requirements.txt
boto3~=1.17.28
psycopg2~=2.8.6
botocore~=1.20.36
python-dateutil~=2.8.1